factor out menu
